requirements.txt
flask==3.1.3
pyyaml==6.0.3
gunicorn==25.1.0